+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 
+ * Add `RealSenseFrameGrabber` and `OpenKinect2FrameGrabber` to `FrameGrabber.list` to have them loaded by default
+ * Remove confusing and no longer useful profiles from the `pom.xml` file
  * Provide new `FFmpegFrameGrabber(InputStream)` and `FFmpegFrameRecorder(OutputStream)` constructors ([issue #95](https://github.com/bytedeco/javacv/issues/95))
  * Make `FrameFilter`, `FrameGrabber`, and `FrameRecorder` implement `Closeable` to let us try-with-resources
  * Fix potential crash when recording audio with `FFmpegFrameRecorder`